requirements.txt
absl-py==0.9.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.4.2
future==0.18.2
google-auth==1.11.2
google-auth-oauthlib==0.4.1
grpcio==1.27.2
idna==2.9
imageio==2.8.0
imageio-ffmpeg==0.4.1
joblib==0.14.1
kiwisolver==1.1.0
Markdown==3.2.1
matplotlib==3.2.0
moviepy==1.0.1
mpmath==1.1.0
nibabel==3.0.1
numpy==1.18.1
oauthlib==3.1.0
pandas==1.0.1
Pillow==7.0.0
pkg-resources==0.0.0
proglog==0.1.9
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
python-dateutil==2.8.1
pytorch-ignite==0.3.0
pytorch-lightning==0.7.1
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
SimpleITK==1.2.4
six==1.14.0
tensorboard==2.1.1
tensorboardX==2.0
torch==1.4.0
torch-dct==0.1.5
torchio==0.13.9
torchsummary==1.5.1
torchvision==0.5.0
tqdm==4.43.0
urllib3==1.25.8
Werkzeug==1.0.0
